+++ duneanaobj/StandardRecord/SRPFP.h
@@ -9,11 +9,11 @@
 #define DUNEANAOBJ_SRPFP_H
 
 #include "duneanaobj/StandardRecord/SRVector3D.h"
-#include "duneanaobj/StandardRecord/SRTrueParticle.h"
+#include "duneanaobj/StandardRecord/SRRecoObjBase.h"
 
 namespace caf
 {
-  class SRPFP
+  class SRPFP : public SRRecoObjBase
   {
     public:
       // less typing further below