+++ duneanaobj/StandardRecord/SRRecoParticle.h
@@ -13,6 +13,8 @@
 
 namespace caf
 {
+  class SRRecoObjBase;
+
   /// \brief Reconstructed particle candidate
   class SRRecoParticle
   {
@@ -49,7 +51,8 @@ namespace caf
 
       std::vector<TrueParticleID> truth;              ///< Associated SRTrueParticle(s), if relevant (use SRTruthBranch::Particle() with these IDs to grab them)
       std::vector<float>   truthOverlap;              ///< Fractional overlap between this reco particle and true particle
-
+      
+      SRRecoBaseID recoobj;                           ///< Id of the reconstructed object this particle is built on
   };
 
 } // caf